some refinements to what Rob did
